backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.20
cloudinary==1.41.0
redis==5.2.1
exponent-server-sdk==2.1.0
slowapi==0.1.9
httpx==0.28.1
eval_type_backport==0.2.0
pydantic[email]
